+++ server/lib/python/cartodb_services/cartodb_services/refactor/config/__init__.py
@@ -1,0 +1,2 @@
+from rate_limits import RateLimitsConfig, RateLimitsConfigBuilder
+from legacy_rate_limits import RateLimitsConfigLegacyBuilder
