+++ server/lib/python/cartodb_services/cartodb_services/tools/__init__.py
@@ -3,5 +3,5 @@ from coordinates import Coordinate
 from polyline import PolyLine
 from log import Logger, LoggerConfig
 from rate_limiter import RateLimiter
-from exceptions import RateLimitExceeded
-from service_manager import ServiceManager, LegacyServiceManager
+from service_manager import ServiceManager, RateLimitExceeded
+from legacy_service_manager import  LegacyServiceManager
