--- server/lib/python/cartodb_services/cartodb_services/tools/exceptions.py
@@ -1,3 +1,0 @@
-class RateLimitExceeded(Exception):
-    def __str__(self):
-            return repr('Rate limit exceeded')
